+++ deco.php
@@ -1,6 +1,4 @@
 <?php
-
-//page déconnexion pour mettre fin à la session : 
 session_start();                                         //on démarre la session pour pouvoir utiliser les variables de session
 unset($_SESSION['email']);                               //on détruit la variable de session
 
@@ -9,6 +7,6 @@ if (ini_get("session.use_cookies")) {                    //on vérifie si les co
 }
 
 session_destroy();                                       //on détruit la session
-header('location: LoginPage.php');                       //on redirige vers la page de connexion
+header('location: index.php');                       //on redirige vers la page de connexion
 
 ?>
